+++ example/bin.dart
@@ -6,8 +6,9 @@ void main() {
   final rand = Random(12345);
 
   final extents = List.generate(11, (i) => i * 10).edgesToExtents();
-  final data = extents.rands(20, source: rand)!;
-  print(extents.computeBins(data));
-  print(extents.computeCounts(data));
-  print(extents.computeHistogram(data));
+  final data = extents.rands(20, source: rand)!.toList();
+
+  print(extents.computeBins(data).asTable());
+  print(extents.computeCounts(data).asTable());
+  print(extents.computeHistogram(data).asTable());
 }